requirements.txt
Django==1.10.0
gunicorn==19.6.0
djangorestframework==3.7.7